pyproject.toml
[tool.poetry]
name = "dao-records"
version = "0.1.0"
description = ""
authors = ["WATUNeed <maxconal228@gmail.com>"]
readme = "README.md"
packages = [{include = "dao_records"}]

[tool.poetry.dependencies]
python = "^3.10"
eth-brownie = "^1.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
